requirements.txt
PySide2
QT-PyQt-PySide-Custom-Widgets